# backend/app/api/routes/auth.py
from fastapi import APIRouter, Depends, HTTPException, status
from sqlmodel import Session

from app.api.deps import get_current_user, get_db
from app.db.models import User
from app.dto.auth import AccountUpdateRequest, LoginRequest, RegisterRequest, TokenResponse, UserMeResponse
from app.services.auth_service import AuthService

router = APIRouter()


@router.post("/register", response_model=UserMeResponse, status_code=status.HTTP_201_CREATED)
def register(payload: RegisterRequest, session: Session = Depends(get_db)) -> UserMeResponse:
    service = AuthService(session)
    user = service.register(payload)
    return UserMeResponse.model_validate(user)


@router.post("/login", response_model=TokenResponse)
def login(payload: LoginRequest, session: Session = Depends(get_db)) -> TokenResponse:
    service = AuthService(session)
    token = service.login(payload)
    if not token:
        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED, detail="用户名或密码错误")
    return token


@router.get("/me", response_model=UserMeResponse)
def me(current_user: User = Depends(get_current_user)) -> UserMeResponse:
    return UserMeResponse.model_validate(current_user)


@router.put("/me", response_model=UserMeResponse)
def update_me(
    payload: AccountUpdateRequest,
    current_user: User = Depends(get_current_user),
    session: Session = Depends(get_db),
) -> UserMeResponse:
    user = AuthService(session).update_account(current_user, payload)
    return UserMeResponse.model_validate(user)